Conflicts:
	projects/OG-Bloomberg/src/com/opengamma/bbg/BloombergSecuritySource.java
	projects/OG-Bloomberg/src/com/opengamma/bbg/loader/BloombergHistoricalTimeSeriesLoader.java
	projects/OG-Bloomberg/tests/unit/com/opengamma/bbg/RemoteBloombergSecuritySourceWithActiveMQTest.java
	projects/OG-Bloomberg/tests/unit/com/opengamma/bbg/security/BloombergSecurityProviderTest.java
	projects/OG-BloombergExample/config/fullstack/fullstack-bloombergexample-viewprocessor-spring.xml
	projects/OG-BloombergExample/src/com/opengamma/bloombergexample/loader/ExampleTimeSeriesRatingLoader.java
	projects/OG-Component/src/com/opengamma/component/factory/web/WebsiteViewportsComponentFactory.java
	projects/OG-Core/src/com/opengamma/core/config/impl/EHCachingConfigSource.java
	projects/OG-Engine/src/com/opengamma/engine/marketdata/historical/AbstractHistoricalMarketDataProvider.java
	projects/OG-Examples/src/com/opengamma/examples/loader/ExampleTimeSeriesRatingLoader.java
	projects/OG-Examples/src/com/opengamma/examples/tool/ExampleDatabasePopulator.java
	projects/OG-Financial/src/com/opengamma/financial/analytics/forwardcurve/FXForwardCurveConfigPopulator.java
	projects/OG-Integration/src/com/opengamma/integration/copier/portfolio/rowparser/ExchangeTradedRowParser.java
	projects/OG-Integration/src/com/opengamma/integration/copier/portfolio/writer/MasterPortfolioWriter.java
	projects/OG-Language/Client/src/com/opengamma/language/view/ViewIdFunction.java
	projects/OG-Master/src/com/opengamma/master/config/impl/InMemoryConfigMaster.java
	projects/OG-Master/src/com/opengamma/master/exchange/impl/InMemoryExchangeMaster.java
	projects/OG-Master/src/com/opengamma/master/historicaltimeseries/impl/MasterHistoricalTimeSeriesSource.java
	projects/OG-Master/src/com/opengamma/master/marketdatasnapshot/impl/InMemorySnapshotMaster.java
	projects/OG-Master/src/com/opengamma/master/position/impl/InMemoryPositionMaster.java
	projects/OG-Master/src/com/opengamma/master/region/RegionSearchResult.java
	projects/OG-Master/src/com/opengamma/master/user/UserDocument.java
	projects/OG-Master/src/com/opengamma/master/user/UserSearchRequest.java
	projects/OG-Master/src/com/opengamma/master/user/UserSearchResult.java
	projects/OG-Master/src/com/opengamma/master/user/impl/InMemoryUserMaster.java
	projects/OG-Master/src/com/opengamma/master/user/impl/MasterUserSource.java
	projects/OG-MasterDB/src/com/opengamma/masterdb/config/DbConfigMaster.java
	projects/OG-MasterDB/src/com/opengamma/masterdb/config/DbConfigWorker.java
	projects/OG-MasterDB/src/com/opengamma/masterdb/historicaltimeseries/DbHistoricalTimeSeriesMaster.java
	projects/OG-MasterDB/src/com/opengamma/masterdb/user/DbUserMaster.java
	projects/OG-MasterDB/tests/unit/com/opengamma/masterdb/user/DbUserMasterTest.java
	projects/OG-Util/src/com/opengamma/util/ehcache/EHCacheUtils.java
	projects/OG-Web/src/com/opengamma/web/security/WebSecurityResource.java
	projects/OG-Web/src/com/opengamma/web/server/push/grid/PushLiveResultsService.java
	projects/OG-Web/tests/unit/com/opengamma/web/server/push/RestEntitySubscriptionTest.java
	projects/OG-Web/tests/unit/com/opengamma/web/server/push/ViewDefinitionEntriesResourceTest.java

+++ projects/OG-Analytics/src/com/opengamma/financial/convention/daycount/ThirtyEThreeSixtyISDA.java
@@ -45,12 +45,12 @@ public class ThirtyEThreeSixtyISDA extends ThirtyThreeSixtyTypeDayCount {
 
   public double getDayCountFraction(final LocalDate firstDate, final LocalDate secondDate, final boolean isMaturity) {
     testDates(firstDate, secondDate);
-    double d1 = firstDate.getDayOfMonth();
-    double d2 = secondDate.getDayOfMonth();
-    final double m1 = firstDate.getMonthOfYear().getValue();
-    final double m2 = secondDate.getMonthOfYear().getValue();
-    final double y1 = firstDate.getYear();
-    final double y2 = secondDate.getYear();
+    int d1 = firstDate.getDayOfMonth();
+    int d2 = secondDate.getDayOfMonth();
+    final int m1 = firstDate.getMonthOfYear().getValue();
+    final int m2 = secondDate.getMonthOfYear().getValue();
+    final int y1 = firstDate.getYear();
+    final int y2 = secondDate.getYear();
     if (d1 == firstDate.getMonthOfYear().getLastDayOfMonth(firstDate.isLeapYear())) {
       d1 = 30;
     }
